# Conflicts:
#	WooCommerce/src/main/res/layout/fragment_card_reader_payment.xml
#	WooCommerce/src/main/res/values/dimens_payments.xml

+++ WooCommerce/src/main/kotlin/com/woocommerce/android/model/Refund.kt
@@ -63,28 +63,28 @@ data class Refund(
 
 fun WCRefundModel.toAppModel(): Refund {
     return Refund(
-            id,
-            dateCreated,
-            amount.roundError(),
-            reason,
-            automaticGatewayRefund,
-            items.map { it.toAppModel() },
-            shippingLineItems.map { it.toAppModel() }
+        id,
+        dateCreated,
+        amount.roundError(),
+        reason,
+        automaticGatewayRefund,
+        items.map { it.toAppModel() },
+        shippingLineItems.map { it.toAppModel() }
     )
 }
 
 fun WCRefundItem.toAppModel(): Refund.Item {
     return Refund.Item(
-            productId ?: -1,
-            -quantity, // WCRefundItem.quantity is NEGATIVE
-            itemId,
-            name ?: "",
-            variationId ?: -1,
-            -subtotal.roundError(), // WCRefundItem.subtotal is NEGATIVE
-            -(total ?: BigDecimal.ZERO).roundError(), // WCRefundItem.total is NEGATIVE
-            -totalTax.roundError(), // WCRefundItem.totalTax is NEGATIVE
-            sku ?: "",
-            price?.roundError() ?: BigDecimal.ZERO,
+        productId ?: -1,
+        -quantity, // WCRefundItem.quantity is NEGATIVE
+        itemId,
+        name ?: "",
+        variationId ?: -1,
+        -subtotal.roundError(), // WCRefundItem.subtotal is NEGATIVE
+        -(total ?: BigDecimal.ZERO).roundError(), // WCRefundItem.total is NEGATIVE
+        -totalTax.roundError(), // WCRefundItem.totalTax is NEGATIVE
+        sku ?: "",
+        price?.roundError() ?: BigDecimal.ZERO,
         metaData?.get(0)?.value?.toString()?.toLongOrNull() ?: -1
     )
 }
@@ -125,8 +125,8 @@ fun List<Refund>.getNonRefundedProducts(
  */
 fun List<Refund>.getMaxRefundQuantities(
     products: List<Order.Item>
-): Map<Long, Int> {
-    val map = mutableMapOf<Long, Int>()
+): Map<Long, Float> {
+    val map = mutableMapOf<Long, Float>()
     val groupedRefunds = this.flatMap { it.items }.groupBy { it.orderItemId }
     products.map { item ->
         map[item.itemId] = item.quantity - (groupedRefunds[item.itemId]?.sumBy { it.quantity } ?: 0)
